Attempts (Should Have)
The user is allowed to make up to 3 mistakes at each level.
If a user makes 3 mistakes, they’re giving the following options
restart the same level
go back to any of the previous levels
switch to the latest level with another algorithm  (updated on Jan. 11th at 10:45 am)
or quit the game